Initial implementation of gRPC endpoint.
