+++ cmd/accountd/grpc/users/service.proto
@@ -1,0 +1,45 @@
+// Copyright (c) 2020 Richard Youngkin. All rights reserved.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
+syntax = "proto3";
+package users;
+
+import "google/protobuf/empty.proto";
+
+// The package name for the generated go file. Not sure exactly what ".;users" except
+// that this will put a 'package users' line in the generated file.
+option go_package = ".;users";
+
+service User {
+    rpc GetUser (GetUserRqst) returns (UserResponse) {}
+    rpc GetUsers(google.protobuf.Empty) returns (UsersResponse) {}
+}
+
+enum RoleType {
+    PRIMARY = 0;
+    UNRESTRICTED = 1;
+    RESTRICTED = 2;
+}
+    
+message UserResponse {
+    int64    AccountID = 1;
+    string   HREF = 2;    
+    int64    ID = 3;   
+    string   Name = 4;   
+    string   EMail = 5;    
+    RoleType Role = 6;
+    string   Password  = 7;
+}
+
+message UsersResponse {
+    repeated UserResponse users = 1;
+}
+
+message GetUserRqst {
+    int64 id = 1;
+}
+
+message GetUsersRqst {
+    google.protobuf.Empty empty = 1;
+}
