+++ cmd/accountd/http/doc.go
@@ -1,0 +1,10 @@
+// Copyright (c) 2020 Richard Youngkin. All rights reserved.
+// Use of this source code is governed by a MIT-style
+// license that can be found in the LICENSE file.
+
+/*
+Package handlers contains the implementations of the HTTP handlers for the 'accountd' service. There
+are several resources that are part of an account. These include the users that are part of an account,
+and the account that has one or more associated users.
+*/
+package handlers
